+++ master.cfg
@@ -9,10 +9,11 @@ from buildbot.plugins import worker, steps, changes, util, reporters
 
 from ursabot.hooks import GithubHook
 from ursabot.utils import Config, ConfigError, Collection
-from ursabot.workers import (DockerLatentWorker, BuilderConfig, TryScheduler,
-                             AnyBranchScheduler, ForceScheduler)
-from ursabot.factories import (BuildFactory, cpp, python, cpp_conda,
-                               python_conda, ursabot_test)
+from ursabot.workers import DockerLatentWorker
+from ursabot.reporters import ZulipMailNotifier
+from ursabot.builders import (BuildFactory, cpp, python, Builder,
+                              python_conda, UrsabotTest, ArrowCppCondaTest)
+from ursabot.schedulers import AnyBranchScheduler, TryScheduler, ForceScheduler
 from ursabot.docker import arrow_images, ursabot_images
 
 
@@ -49,7 +50,7 @@ c = BuildmasterConfig = {
     }
 }
 
-############################# HOOKS & REPORTERS ###############################
+############################## AUTHENTICATION ################################
 
 if conf.github.auth.enabled:
     c['www']['auth'] = util.GitHubAuth(
@@ -76,6 +77,8 @@ if conf.github.auth.enabled:
         ]
     )
 
+################################ HOOKS ########################################
+
 if conf.github.event_hook.enabled:
     c['change_hook_dialects'] = {
         'github': {
@@ -86,7 +89,9 @@ if conf.github.event_hook.enabled:
         }
     }
 
-if conf.github.status_push.enabled:
+################################ REPORTERS ####################################
+
+if conf.reporters.github_status.enabled:
     c['services'].append(
         reporters.GitHubStatusPush(
             token=conf.github.token,
@@ -94,6 +99,14 @@ if conf.github.status_push.enabled:
             context=util.Interpolate('buildbot/%(prop:buildername)s'),
             startDescription='Build started.',
             endDescription='Build done.'
+        )
+    )
+
+if conf.reporters.zulip.enabled:
+    c['services'].append(
+        ZulipMailNotifier(
+            zulipaddr=conf.reporters.zulip.address,
+            fromaddr=conf.reporters.zulip['from']
         )
     )
 
@@ -160,29 +173,19 @@ for worker in conf.workers:
         )
     )
 
-################################ BUILDERS #####################################
-
 workers_by_arch = workers.groupby('arch')
 
-arrow_builders = Collection([
-    BuilderConfig(
-        name=image.repo,
-        factory=cpp_conda,
-        workers=workers_by_arch['amd64'],
-        properties={'docker_image': str(image)}
-    ) for image in arrow_images.filter(name='cpp', variant='conda',
-                                       tag='worker')
-])
+################################ BUILDERS #####################################
 
-ursabot_builders = Collection([
-    BuilderConfig(
-        name='ursabot-test',
-        tags=['docker', 'python'],
-        factory=ursabot_test,
-        workers=workers_by_arch['amd64'],
-        properties={'docker_image': str(image)}
-    ) for image in ursabot_images.filter(tag='worker')
-])
+arrow_builders = [
+    ArrowCppCondaTest(image=image, workers=workers_by_arch['amd64'])
+    for image in arrow_images.filter(name='cpp', variant='conda', tag='worker')
+]
+
+ursabot_builders = [
+    UrsabotTest(image=image, workers=workers_by_arch['amd64'])
+    for image in ursabot_images.filter(tag='worker')
+]
 
 ################################ SCHEDULERS ###################################
 
@@ -333,36 +336,3 @@ if conf.projects.ursabot.enabled:
 #         arrow_builders.append(b)
 #
 #     c['builders'].extend(arrow_builders)
-
-
-# trigger
-
-####### BUILDBOT SERVICES
-
-# 'services' is a list of BuildbotService items like reporter targets. The
-# status of each build will be pushed to these targets. buildbot/reporters/*.py
-# has a variety to choose from, like IRC bots.
-
-# template=u'''\
-# <h4>Build status: {{ summary }}</h4>
-# <p> Worker used: {{ workername }}</p>
-# {% for step in build['steps'] %}
-# <p> {{ step['name'] }}: {{ step['result'] }}</p>
-# {% endfor %}
-# <p><b> -- The Buildbot</b></p>
-# '''
-
-# if ZULIP:
-#     c['services'].append(
-#         reporters.MailNotifier(
-#             fromaddr=EMAIL,
-#             sendToInterestedUsers=False,
-#             extraRecipients=[ZULIP],
-#             messageFormatter=reporters.MessageFormatter(
-#                 template=template,
-#                 template_type='html',
-#                 wantProperties=True,
-#                 wantSteps=True
-#             )
-#         )
-#     )
