+++ ursabot/reporters.py
@@ -1,0 +1,25 @@
+from buildbot.plugins import reporters
+
+
+_template = u'''\
+<h4>Build status: {{ summary }}</h4>
+<p> Worker used: {{ workername }}</p>
+{% for step in build['steps'] %}
+<p> {{ step['name'] }}: {{ step['result'] }}</p>
+{% endfor %}
+<p><b> -- The Buildbot</b></p>
+'''
+
+
+class ZulipMailNotifier(reporters.MailNotifier):
+
+    def __init__(self, zulipaddr, fromaddr, template=None):
+        formatter = reporters.MessageFormatter(
+            template=template or _template,
+            template_type='html',
+            wantProperties=True,
+            wantSteps=True
+        )
+        super().__init__(fromaddr=fromaddr, extraRecipients=[zulipaddr],
+                         messageFormatter=formatter,
+                         sendToInterestedUsers=False)
