+++ ursabot/schedulers.py
@@ -1,0 +1,20 @@
+from buildbot.plugins import schedulers
+
+
+class SchedulerMixin:
+
+    def __init__(self, *args, builders, **kwargs):
+        builder_names = [b.name for b in builders]
+        super().__init__(*args, builderNames=builder_names, **kwargs)
+
+
+class ForceScheduler(SchedulerMixin, schedulers.ForceScheduler):
+    pass
+
+
+class TryScheduler(SchedulerMixin, schedulers.Try_Userpass):
+    pass
+
+
+class AnyBranchScheduler(SchedulerMixin, schedulers.AnyBranchScheduler):
+    pass
